Update README.md
Adding  a link to a MySQL store created for a project I am working on.  I am working with a partner to replicate  your session store tests in the my project (trying to get him involved in the opensource community), but it seems to be working consistently for us.  I am not sure what the protocol is for announcing this related project - so I am hoping this does it.  Thanks Todd